+++ build.ps1
@@ -1,5 +1,5 @@
 param(
-    [string]$Version  = "v2.2.0",
+    [string]$Version  = "v2.3",
     [string]$OutDir   = "builds"
 )